refactor: Split make
